dodanie obsługi przycisku usun lot oraz zmiana relacji miedzy encjami Airport i Flight do wiele do jednego

+++ src/main/java/com/qbaaa/flights/Controller/QuestionFlightController.java
@@ -5,13 +5,13 @@ import com.qbaaa.flights.Model.*;
 import com.qbaaa.flights.Repository.AirportRepository;
 import com.qbaaa.flights.Repository.FlightCargoRepository;
 import com.qbaaa.flights.Repository.FlightPersonalRepository;
+import com.qbaaa.flights.Repository.FlightRepository;
 import com.qbaaa.flights.Request.FlightForm;
 import com.qbaaa.flights.Response.MessageResponse;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-
 import java.util.List;
 
 @CrossOrigin(origins = "*")
@@ -26,6 +26,9 @@ public class QuestionFlightController
 
     @Autowired
     FlightCargoRepository flightCargoRepository;
+
+    @Autowired
+    FlightRepository flightRepository;
 
     @PostMapping("/addAirport")
     public ResponseEntity<?> addAirport(@RequestBody Airport addAirport)
@@ -43,10 +46,10 @@ public class QuestionFlightController
     @GetMapping("/getAirport")
     public ResponseEntity<?> getAirport()
     {
-        List listAirport = airportRepository.findAll();
+        List<Airport> listAirport = airportRepository.findAll();
 
         if(listAirport.isEmpty())
-            return new ResponseEntity<>(new MessageResponse("Barak lotnisk w bazie"),HttpStatus.BAD_REQUEST);
+            return new ResponseEntity<>(new MessageResponse("Brak lotnisk w bazie"),HttpStatus.BAD_REQUEST);
         else
             return new ResponseEntity<>(listAirport, HttpStatus.OK);
 
@@ -56,11 +59,12 @@ public class QuestionFlightController
     public ResponseEntity<?> deleteAirport(@PathVariable("id") Long id)
     {
         if (id == 0)
-        {
             return new ResponseEntity<>(new MessageResponse("Nie wybrano żadnego lotniska!") ,HttpStatus.BAD_REQUEST);
-        }
 
         Airport delAirport = airportRepository.findById(id).orElseThrow(() -> new RuntimeException("Nie ma lotniska o zadanym id!"));
+
+        if (!delAirport.getEndA().isEmpty() || !delAirport.getStartA().isEmpty())
+            return new ResponseEntity<>(new MessageResponse("NIE mozna usunąc lotniska, poniewaz jest powiazany z lotam."),HttpStatus.BAD_REQUEST);
 
         airportRepository.delete(delAirport);
 
@@ -80,10 +84,34 @@ public class QuestionFlightController
         else
         {
             flightCargoRepository.save(new FlightCargo(start, end, form.getTimeStart(), form.getDurationTimeMinutes(), form.getFeeConst(), form.getFeeForKiloPack()));
-            System.out.println(new FlightCargo(start, end, form.getTimeStart(), form.getDurationTimeMinutes(), form.getFeeConst(), form.getFeeForKiloPack()));
         }
 
         return new ResponseEntity<>(new MessageResponse("Dodano pomyslnie lot."),HttpStatus.OK);
+    }
+
+    @GetMapping("/getFlight")
+    public ResponseEntity<?> getFlight()
+    {
+        List<Flight> list = flightRepository.findAll();
+
+        if (list.isEmpty())
+            return new ResponseEntity<>(new MessageResponse("Brak lotow w bazie !!!"), HttpStatus.BAD_REQUEST);
+        else
+            return new ResponseEntity<>(list, HttpStatus.OK);
+
+    }
+
+    @DeleteMapping("/deleteFlight/{id}")
+    public ResponseEntity<?> deleteFlight(@PathVariable Long id)
+    {
+        if (id == 0)
+            return new ResponseEntity<>(new MessageResponse("Nie wybrano żadnego lotu!!!") ,HttpStatus.BAD_REQUEST);
+
+        Flight del = flightRepository.findById(id).orElseThrow(() -> new RuntimeException("Nie ma lotu o zadanym id!"));
+
+        flightRepository.delete(del);
+
+        return new ResponseEntity<>(new MessageResponse("Usuniecie lotu przebiegło pomyślnie."),HttpStatus.OK);
     }
 
 }
